webserver/requirements.txt
strands-agents
strands-agents-tools
strands-agents-builder
flask
watchdog
strands-agents[ollama]